Move specs.py to specs module.
