+++ metricflow/dataflow/builder/dataflow_plan_builder.py
@@ -51,7 +51,7 @@ from metricflow.plan_conversion.column_resolver import DefaultColumnAssociationR
 from metricflow.plan_conversion.node_processor import PreDimensionJoinNodeProcessor
 from metricflow.plan_conversion.sql_dataset import SqlDataSet
 from metricflow.plan_conversion.time_spine import TimeSpineSource
-from metricflow.specs import (
+from metricflow.specs.specs import (
     MetricSpec,
     MetricInputMeasureSpec,
     LinkableInstanceSpec,
