+++ metricflow/dataset/convert_semantic_model.py
@@ -20,7 +20,7 @@ from dbt_semantic_interfaces.objects.elements.dimension import Dimension, Dimens
 from dbt_semantic_interfaces.objects.elements.entity import Entity
 from dbt_semantic_interfaces.objects.elements.measure import Measure
 from metricflow.model.spec_converters import MeasureConverter
-from metricflow.specs import (
+from metricflow.specs.specs import (
     TimeDimensionSpec,
     DimensionSpec,
     EntitySpec,
