+++ metricflow/dataset/dataset.py
@@ -6,7 +6,7 @@ from typing import Sequence
 from dbt_semantic_interfaces.references import TimeDimensionReference
 from metricflow.instances import InstanceSet, TimeDimensionInstance
 from dbt_semantic_interfaces.validations.unique_valid_name import MetricFlowReservedKeywords
-from metricflow.specs import TimeDimensionSpec
+from metricflow.specs.specs import TimeDimensionSpec
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
