+++ metricflow/spec_set_transforms.py
@@ -1,6 +1,6 @@
 from typing import Set
 
-from metricflow.specs import InstanceSpecSetTransform, InstanceSpecSet
+from metricflow.specs.specs import InstanceSpecSetTransform, InstanceSpecSet
 
 
 class ToElementNameSet(InstanceSpecSetTransform[Set[str]]):
