+++ metricflow/test/dataflow/builder/test_measure_additiveness.py
@@ -1,6 +1,6 @@
 from metricflow.dataflow.builder.measure_additiveness import group_measure_specs_by_additiveness
 from dbt_semantic_interfaces.objects.aggregation_type import AggregationType
-from metricflow.specs import MeasureSpec, NonAdditiveDimensionSpec
+from metricflow.specs.specs import MeasureSpec, NonAdditiveDimensionSpec
 
 
 def test_bucket_measure_specs_by_additiveness() -> None:  # noqa: D
