+++ metricflow/test/model/test_where_filter_spec.py
@@ -3,7 +3,7 @@ import logging
 from dbt_semantic_interfaces.objects.filters.where_filter import WhereFilter
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.references import EntityReference
-from metricflow.specs import (
+from metricflow.specs.specs import (
     ColumnAssociationResolver,
     LinkableSpecSet,
     DimensionSpec,
