+++ metricflow/test/query/test_query_parser.py
@@ -8,7 +8,7 @@ from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
 from metricflow.filters.time_constraint import TimeRangeConstraint
 from metricflow.errors.errors import UnableToSatisfyQueryError
 from metricflow.plan_conversion.time_spine import TimeSpineSource
-from metricflow.specs import (
+from metricflow.specs.specs import (
     MetricSpec,
     DimensionSpec,
     TimeDimensionSpec,
